pyproject.toml
[tool.poetry]
name = "football-mini-app"
version = "0.1.0"
description = "A football mini app backend using FastAPI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
tortoise-orm = "0.18.0"
fastapi-admin = "*"
aiosqlite = "*"
pydantic-settings = "*"
python-dotenv = "*"
pyjwt = "*"
passlib = {extras = ["bcrypt"], version = "*"}
fastapi-cors = "*"
django = "^5.2.4"
djangorestframework = "^3.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 